Ask user to confirm if not all documents are supported
Issues appear because unlike the file picker, we can't control the files
that the user drags. We can only control the ones dropped. There are two
main edge-cases:
  1. When none of the dropped files are supported
  2. When some of the dropped files are supported

Ignoring in case 1. seems to make sense. It's already what we do when
the user drops text.

For case 2. it shows the user a prompt to confirm that the user wants to
continue even though some files were ignored.